+++ examples/nodejs/plt/v1/transfer.ts
@@ -1,17 +1,10 @@
-import {
-    AccountAddress,
-    AccountTransactionType,
-    buildAccountSigner,
-    parseWallet,
-    serializeAccountTransactionPayload,
-} from '@concordium/web-sdk';
+import { AccountAddress, AccountTransactionType, serializeAccountTransactionPayload } from '@concordium/web-sdk';
 import { ConcordiumGRPCNodeClient } from '@concordium/web-sdk/nodejs';
 import { CborMemo, TokenAmount, TokenId, V1 } from '@concordium/web-sdk/plt';
 import { credentials } from '@grpc/grpc-js';
 import meow from 'meow';
-import { readFileSync } from 'node:fs';
 
-import { parseEndpoint } from '../../shared/util.js';
+import { parseEndpoint, parseKeysFile } from '../../shared/util.js';
 
 const cli = meow(
     `
@@ -84,17 +77,17 @@ const client = new ConcordiumGRPCNodeClient(address, Number(port), credentials.c
     console.log('Specified transfer:', JSON.stringify(transfer, null, 2));
 
     if (cli.flags.walletFile !== undefined) {
-        // Read wallet-file
-        const walletFile = readFileSync(cli.flags.walletFile, 'utf8');
-        const wallet = parseWallet(walletFile);
-        const sender = AccountAddress.fromBase58(wallet.value.address);
-        const signer = buildAccountSigner(wallet);
+        const [sender, signer] = parseKeysFile(cli.flags.walletFile);
 
         // From a service perspective:
         // create the token instance
-        const token = await V1.Token.fromId(client, tokenSymbol);
-        const transaction = await V1.Token.transfer(token, sender, transfer, signer);
-        console.log(`Transaction submitted with hash: ${transaction}`);
+        try {
+            const token = await V1.Token.fromId(client, tokenSymbol);
+            const transaction = await V1.Token.transfer(token, sender, transfer, signer);
+            console.log(`Transaction submitted with hash: ${transaction}`);
+        } catch (e) {
+            console.error(e);
+        }
     } else {
         // Or from a wallet perspective:
         // Create transfer payload