+++ packages/sdk/src/plt/v1/Token.ts
@@ -59,7 +59,7 @@ export class NotAllowedError extends TokenError {
 
     /**
      * Constructs a new NotAllowedError.
-     * @param {AccountAddress.Type} address - The account address of the receiver.
+     * @param {AccountAddress.Type} receiver - The account address of the receiver.
      */
     constructor(public readonly receiver: AccountAddress.Type) {
         super(
@@ -246,9 +246,13 @@ export async function transfer(
     sender: AccountAddress.Type,
     payload: TokenTransfer | [TokenTransfer],
     signer: AccountSigner,
-    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5)
+    expiry: TransactionExpiry.Type = TransactionExpiry.futureMinutes(5),
+    skipValidation = true
 ): Promise<TransactionHash.Type> {
-    await validateTransfer(token, sender, payload);
+    // TODO: this should be removed when the validation function does not fail due to unimplemented grpc endpoints
+    if (!skipValidation) {
+        await validateTransfer(token, sender, payload);
+    }
 
     const ops: TokenHolderOperation[] = [payload].flat().map((p) => ({ [TokenOperationType.Transfer]: p }));
     const encoded = createTokenHolderPayload(token.info.id, ops);